int32

+++ linbox/examples/smith.C
@@ -142,7 +142,7 @@ int main(int argc, char* argv[])
 	}
 	else if (algo == "ilio") {
 
-		PIR R(m);
+                PIR R( (int32_t)m);
 
 		BlasMatrix<PIR> M(R);
